refactoring plot
